Move UserSegment from mysql dialect to common dcl segment
